requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
kaggle
tqdm
matplotlib
Pillow
albumentations
transformers